tests: Disable reconnection in StompTests.test_echo
This test fails randomly on the CI, however since we added reconnect
support, when this tests fail, it starts an endless reconnect loop,
blocking the tests forever.

This an attempt to use the nr_retries and reconnect_interval to prevent
the reconnection attempts. I hope that this will make the client fail
after the timeout.

To add the new arguments, I cleaned up the code a little bit.

Change-Id: Ib1d5bebd939b11c4860a819946ea4102fdea82c2
Signed-off-by: Nir Soffer <nsoffer@redhat.com>