logUtils: Add threaded logging handler
ThreadedHandler queues log records and handle them in a background
thread using UserGroupEnforcingHandler logfile handler. This prevents
blocking in various threads if writing  to the log file becomes slow.

Blocking when writing to log file may cause vdsm to miss heartbeats and
make VM and storage domain non responsive. This may lead to unneeded
soft fencing or moving host to non-responsive or non-operational state,
leading to migration storms.

If writing to log file is slow for long time, oldest log messages are
dropped. The size of the queue is configurable using logger.conf, but
for easy migration we use a default value.

Python 3 introduced a QueueHandler and QueueListener, providing similar
functionality. However they not support limiting log queue size in a
useful way; to limit the queue size one has to handle Queue.Full
exception in all loggers, which is not very useful, and will drop new
logs instead of old logs. It also seems impossible to configure
QueueHandler and the QueueListener via logger.conf.

Change-Id: I61ccb2be109131d3bab99889d239a50e403f6c9b
Bug-Url: https://bugzilla.redhat.com/1516831
Signed-off-by: Nir Soffer <nsoffer@redhat.com>