lvm: Support full device path
Many public functions accept a device WWID and create full path to the
device, assuming that the device is a multipath device. We like to allow
creating VG with loop devices for testing.

Change the function creating full device path to accept absolute
paths as is. This allows creating a VG with a loop device path.

Change-Id: Ib60b20f5467c02127467e4bfc2a56dfd42daba63
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>