tests: Start lvm functional tests
The lvm test module provide now a temp_storage fixture that make it easy
to create loop devices and use them to create PVs and VGs. The
temp_storage fixture automatically destroy lvs, vgs and pvs created
during the test.

To allow running lvm tests under mock or inside a container, we install
a special lvmlocal.conf that disable udev. Without this lvm commands
fail waiting for udev events.

The lvm functional tests requires root, so they are skipped if not
running as root. To make it easy to run only these tests as root, they
are also marked as "root".

Here is example run of the functional tests as root (using tox installed
with "pip install --user tox"):

  $ sudo ~/.local/bin/tox -e storage-py27 tests/storage/lvm_test.py -- -m root
  ...

Add first minimal tests:
- create VG from single PV, verify created PV and VG
- create LV, verify created LV

The first 2 tests exposed a bug in the cache in python 3. They are
marked as xfail on python 3 for now.

Change-Id: I7f6b4822f85394fae9083b86ea0c4d817cb36628
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>