managedvolume: Extract _add_volume private helper
The first step in attaching a volume is too complex to inline in
attach_volume. Extract to a helper to make the main logic more clear.

While moving, simplify the flow - instead of checking if the volume
exists and adding it if not, try to add it and get the current volume if
the database fail with VolumeAlreadyExist. This makes the code much more
clear and eliminate the else case.

Change-Id: I9731609148f08e37ddbb2fe148a48ca9559290ea
Signed-off-by: Nir Soffer <nsoffer@redhat.com>