managedvolumedb: Replace sqlite3 with lmbd
Replace the sqlite3 implementation with lmdb, an ultra-fast,
ultra-compact, crash-proof key-value embedded data store.

With this implementation, the concurrency test runs 3 times faster, and
the code is generally simpler.

lmdb must have a single environment per process, and recommends closing
the environment regularly to avoid stale readers. We ensure this by
keeping single environment in the DB class, and closing it when the last
DB instance was closed.

The packages are available on Fedora 28, not sure about RHEL.

Change-Id: Ia8945314796c284f0f48c0f347cc7b2e62bd5e96
Signed-off-by: Nir Soffer <nsoffer@redhat.com>