lvm: Use --select to eliminate errors
We used to reload pvs, vgs and lvs by specifying the entity in the
command line. Here are few examples:

    pvs pv-1 pv-2

    vgs vg-1

    lvs vg-1

    lvs vg-1/lv-1 vg-1/lv-2

If these commands cannot find one of the specified entities, they fail.
When a command fail, we invalidated the filter, and if the new filter is
different, retry the command with the wider filter.

Since we want to retry failing read-only command, these commands would
be retried 10 times in read-only mode. If the entity is missing, all
retries will fail, slowing down the system without any benefit.

To avoid this issue, change reload commands to use --select. When using
select, we can specify which entity we want to query. If the entity does
not exist, the command does not fail. Here are examples of the same
commands using --select:

    pvs --select 'pv_name = pv-1 || pv_name = pv-2'

    vgs --select 'vg_name = vg-1'

    lvs --select 'vg_name = vg-1'

    lvs --select 'vg_name = vg-1 && (lv_name = lv-1 || lv_name = lv-2)'

Using --select introduces a new problem, since we depend on command
failure to refresh a stale filter. Fix this by refreshing the filter
also when a command wants output, but no output was received.

With both changes all integration tests pass without any delay, and we
can enable read-only mode in tests that it was not enable before.

Open issues:

- If a command try to query multiple entities and one of them is
  missing, the command was failing and invalidating the filter. With
  --select, the command will not fail and return data only for the
  existing entities, so we would not refresh a stale filter.  I'm not
  sure if we use commands with multiple entities.

- Need to test the retry mechanism with command that return empty
  output. Currently tests fake only command failures.

- Need to test building a --select pattern.

Change-Id: Iebf2620cc40629ed88c04d0e803092c92d8651cc
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>