storage: External disk support
Support creating a volume backed by external disk. This allows running a
VM from a read-only disk in any format supported by qemu. The external
disk can be a local file or a block device, or a network disk such as
NBD disk.

Interesting use cases for this configuration are:

- Importing VMs from other systems with minimal downtime.
- Preview a backup without copying the data to oVirt volume.
- With virt-v2v, import VMs from foreign systems with minimal downtime.
- Run transient VMs for testing on hosts with limited local storage.

Open issues:

- Supporting local files introduce a security issue, when a user may try
  to access a file or block device on a host from within a VM. We may
  need to limit the location of local external disks. Since this is not
  handled now we log a warning when using external disk.

Change-Id: I3bfa26ff123b716eb64e327c2be380b91aca8496
Signed-off-by: Nir Soffer <nsoffer@redhat.com>