xlease: Support 4k block size
Replace the hard coded BLOCK_SIZE constant with the block_size argument
specified to LeasesVolume() and rebuild_index(). This value is
propagated down to VolumeIndex() and ChangeBlock(), so all writes are
aligned to the actual block size instead of 512 bytes.

Add block_size argument to format_index() since it is needed to write
the metadata area on successful formatting.

With this change, the xlease module is compatible with 4k storage, but
the code formatting the xleases volume must be updated to specify the
block size.

Currently we support only 1MiB alignment. More work is needed to support
2MiB, 4MiB, and 8MiB alignments.

Change-Id: I039cbcaeae73bb0776c2c2ba7f650a6169183091
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Nir Soffer <nsoffer@redhat.com>