tests: Pass alignment value
Change the tmp_vol fixture to include alignment for every user storage,
and pass the alignment to the tests using TemporaryVolume. In the tests,
pass the alignment value to xlease and sanlock calls, and use the
alignment when inspecting storage instead of hard-coded value.

Fixture parameters for 4k storage with 2m, 4m, and 8m alignment were not
added yet since the tests depend on xlease.format_index() and
xlease.VolumeIndex() and they do not accept alignment, and always use
xlease.INDEX_BASE.

Change-Id: I1111a4990629cef1679e81a4d610ea88cffe4254
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Nir Soffer <nsoffer@redhat.com>