tests: Fix gluster tests in python 3
When running the tests on python 3 the test fail with:

>       assert sorted(expected_out) == sorted(ret)
E       TypeError: '<' not supported between instances of 'dict' and 'dict'

Because dicts are not sortable in python 3. I think they were sorted
before by object id which is not very useful.

Fixed by sorting the dicts by the "name" attribute, which seems to be
unique looking in hooks_list.json.

Change-Id: I74001296b4d04e49aee72602be56cca6b7d443b9
Signed-off-by: Nir Soffer <nsoffer@redhat.com>