image: Disable target creation in qemu-img convert
When converting images on file based storage, qemu first truncate the
file to 0, and then truncate it to the final size. This leaves the first
block of the file unallocated, which fails qemu block size detection.
Then qemu fail to write to the image when it try to write data aligned
to 512 bytes on 4k storage.

On block devices creating an image is harmless but pointless, so better
remove it.

Since we always want to use the crate=False, change also all the tests
running real qemu-img to use existing images. This will make it possible
later to run these tests with user storage with both 512 and 4k sector
size.

Issues:

- preallocation does not work when using qemu -n option. This is tracked
  in bug 1738721. We can work around this by preallocating outside
  qemu-img before converting the image. This may cause performance
  regression when converting mostly allocated disk, on legacy
  filesystems that do not support fallocate() (e.g. NFS < 4.2).

Change-Id: Ieb29ef39355c6d043626aab321678925fdc5f244
Bug-Url: https://bugzilla.redhat.com/1592916
Related-To: https://bugzilla.redhat.com/1738657
Related-To: https://bugzilla.redhat.com/1738721
Signed-off-by: Nir Soffer <nsoffer@redhat.com>