fc-scan: Speed up device scanning
Rewrite fc-scan using a thread pool executor, rescanning devices with 10
worker threads instead of the hosts threads.

There is a mutex on each SCSI Host object taken around the scan
operation when rescanning a host, or a target. So for host scanning, we
should use thread per host to get maximum parallelism.

Rescanning an already-discovered device, though, is different and does
not take the mutex.  So there is some potential for parallelism there.

Previously we used the hosts threads to rescan devices. On a setup using
2 fc_hosts we had only 2 threads rescanning devices.  Now we wait until
host scanning is finished, and submit the devices to a new larger thread
pool for device scanning. This should maximize parallelism during device
rescan which is the slow part.

Change-Id: Ib0f74d8765b32943091593dbb0009c794d52efe1
Bug-Url: https://bugzilla.redhat.com/1598266
Signed-off-by: Nir Soffer <nsoffer@redhat.com>