supervdsm: Add run_as_root decorator
We have a lot of code repeating this boilerplate:

    from vdsm.common import supervdsm

    def foo():
        if os.geteuid() != 0:
            return supervdsm.getProxy().module_foo()

        print("I'm root!")

This code can be replace now with:

    from vdsm.common.supervdsm import run_as_root

    @run_as_root
    def foo():
        print("I'm root!")

The decorator assumes that there is an entry point "module_foo" in
supervdsm_api/. This will help to ensure consistency in the code.

For testing the decorator, a new vdsm.test module was added, and the
supervdsm_api/test.py was modified to expose functions in this module.
This serve both as test driver and example code for writing new
supervdsm APIs.

The tests should be run in two ways:

- As root - this is how the tests will run in the CI.

- With running supervdsm, serving the user running tests. This must set
  up manually, as explained in supervdsm_test.py.

Change-Id: I98412125728545718ea621ed9340d22b8be0785c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>