utils: Log both START and FINISH messages in stopwatch
The typical way we use utils.stopwatch is:

    log.debug("Looking up devices")
    with utils.stopwatch("Looked up devices"):
        # time consuming operation...

Enhance stopwatch to log both START and FINISH messages, so we don't
need to remember to add the start log.

We can use now:

    with utils.stopwatch("Looking up devices"):
        # time consuming operation...

And we will have these logs:

    DEBUG [logger] START Looking up devices
    ...
    DEBUG [logger] FINISH Looking up devices: 1.23 seconds

Sometimes we need to format the message. Move formatting into
stopwatch(), using same signature as logging calls, so we can format the
message only if the message is actually logged, and we don't need to
litter the application code with string formatting.

Because this change duplicates existing logs, we need to change the
callers at the same time.

Some of the logs were improved; this should be done in another patch.
Posting the complete change to discuss this approach.

Change-Id: I307ef51bb92592b47266fffcd07b3603d117a678
Signed-off-by: Nir Soffer <nsoffer@redhat.com>