tests: Add tests for lvm.invalidateVG()
Add missing tests for invalidating VG and optionally it's LVs and PVs.

The test for invalidating PVs fails since we reload the VG to find the
VG to find the PV names. This may cause failures if the VG is not ready
when it is invalidated, and prevents using lvm.invalidateVG() in the
tests to test handling of stale items.

Change-Id: I0b58ed3f799b34822bb02ed8249a0aad9433408b
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>