tests: Tests handling of stale cache items
Add missing tests for handling of stale cache items. This simulates the
case when PV, VG, or LV are removed on another host.

The tests handle both kinds of invalidation:
- Dropping entire cache using lvm.invalidateCache()
- Invalidating single VG and optionally its LVs and PVs using
  lvm.invalidateVG()

Older test for using the cache in lvm.getLV() updated to use new
stale_lv() fixture, and is testing now both lvm.invalidateVG() and
lvm.invalidateCache(). The tests for getting all LVS using the cache
is marked as xfail since we don't use the cache yet.

Some tests are marked as xfail since getAllVGs(), getAllPVs(), and
_reloadvgs() may return stubs.

Change-Id: Ifdc4526095ee8f3bee27ada21ff449f9586cf848
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>