common: time: Introduce event_time()
Notifications and VM status use monotonic_time() in milliseconds. The
logic was duplicated in yajssonrpc.Nontfication and vdsm.virt.VM, but
both time values must be synchronized. If the values are not
synchronized, engine may ignore status reports or events, breaking VM
life cycle.

Introduce new time.event_time() removing the duplication and ensuring
that the values are always synchronized.

Change-Id: I4ced82090e1e5ee4050c91954ed5bedab9dbba45
Related-To: https://bugzilla.redhat.com/1828088
Signed-off-by: Nir Soffer <nsoffer@redhat.com>