travis: Disable broken tests on Travis
For some reason, importing sitecustomize fail now on travis:

    import sitecustomize' failed; use -v for traceback

Without sitecustomize system default encoding is "ascii" so the unicode
tests and encoding test in tool_test.py fail.

Adding -v show that sys.setdefaultencoding() is missing when we try to
use it. I verified that importing sitecustomize works on CentOS 7.7
host, so this seems to be a testing environment issue, and we need to
disable these tests until we find understand why it fails.

The pywatch test that fails randomly on Jenkins started to fail now also
on Travis. I marked this tests as xfail also on Travis.

With these changes the tests pass again on Travis.

Change-Id: I3fddb721186108bc0f4ce66def33e631c598759b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>