travis: Expose /dev/log from host
When running sanlock_direct_test.py on travis, sanlock syslog messages
leak to stderr, breaking test output:

    storage/sanlock_direct_test.py sanlock-direct[13001]: init lockspace
    LS:1:/var/tmp/vdsm-storage/mount.file-512/file:0 0x110

This happens because sanlock is using:

    openlog("sanlock-direct", LOG_CONS | LOG_PID, LOG_DAEMON);

And it logs warnings for every "sanlock direct" command:

    syslog(LOG_WARNING, "init lockspace %.48s:%llu:%s:%llu 0x%x", ...

There is no /dev/log in the container, so syslog writes the message to
/dev/console, which seems to be stderr in the container.

This was fixed in sanlock in:

commit 315a0b90612a2b6c3b11d33b48385aa1352fa671
Author: Nir Soffer <nsoffer@redhat.com>
Date:   Tue May 19 09:24:43 2020 -0500

    sanlock: remove LOG_CONS from direct command

But it will take time until this fix is available in all supported
distros.

Mounting host's /dev/log in the container fixes this issue with current
sanlock, logging the messages to host syslog.

Change-Id: I6e3d6aed56987564ab4231a5c372a5a07825c5d7
Signed-off-by: Nir Soffer <nsoffer@redhat.com>