download: Add stress test for vdsm http server
The stress test reproduces random failures when downloading images
concurrently.

Unfortunately this test must run with real setup, and configured
manually with the pool id, domain id, and image id. This is good enough
to reproduce and test the fix.

The test is very slow, about 5 minutes, performing 1000 downloads, but
it reproduces about 160 errors per run without the fix.

Change-Id: I2deb4bfb51099a137f3f8929bba038a912a11a1b
Bug-Url: https://bugzilla.redhat.com/1694972
Signed-off-by: Nir Soffer <nsoffer@redhat.com>