fileSD: Do not use qemu-img preallocation=falloc
We used to defer volume preallocation when creating a raw preallocated
volume by supporting initial_size=1. Then when copying the volume, we
used:

    qemu-img convert ... -o preallocation=falloc src dst

Turns out that this does not work well for NFS < 4.2, and can lead to
sanlock timeouts, storage monitoring failures, and VMs becoming
non-responsive. This is also slower than preallocating volumes using dd
that was done in the past.

To fix this, we drop the idea of deferring preallocation. The initial
size sent by engine is ignored now when creating raw preallocated
volume, and we always preallocate the volume at the creation step.

Creating a raw preallocated volume is done in 2 steps:

1. Create sparse image using qemu-img create, since qemu-img allocates
   the first block by writing zeroes.

2. Allocate the rest of the image using fallocate helper, which was
   changed to use efficient and safe method on any NFS version.

When we copy images, we always use the -n option to use the existing
volume:

    qemu-img convert ... -n src dst

This keeps the target volume preallocated. This was already used for
block storage, and now used also for file storage.

When using create=False dstQcow2Compat and backingFormat are ignored, so
these options were removed.

Change-Id: Ifea5d3dc27e69265d23ba561336d5f00b07151f1
https://bugzilla.redhat.com/1850267
Signed-off-by: Nir Soffer <nsoffer@redhat.com>