fileVolume: Improve logging when creating raw volumes
Log "Request to create RAW volume ..." before creating the volume, for
consistency with other logs.

Log the stopwatch messages for creating the image and preallocating the
volume in INFO level and using FileVolume logger so these logs are
available in the default log level. This will make it easier to diagnose
performance issues with preallocated volumes.

Change-Id: I766f12ae2f66b3eda8a482b8fc23f38a356e4c89
https://bugzilla.redhat.com/1850267
Signed-off-by: Nir Soffer <nsoffer@redhat.com>