lvm: Disable obtain_device_list_from_udev
On RHEL 8.2.1 we see random failures in lvcreate and lvchange when
running tests/storage/stress/reload.py:

    Failed to udev_enumerate_scan_devices.
    Volume group "bz1837199-000000000000000000000-0006" not found.
    Cannot process volume group bz1837199-000000000000000000000-0006

David Teigland suggested[1] to try obtain_device_list_from_udev=0 to
mitigate this. Testing shows that this eliminates the errors, and
improves performance slightly[2].

This is the second time we disable obtain_device_list_from_udev. We did
the same in RHEL 6:

commit 23ce1d87fa98f35d83005e2958ff4065b78ff9d8
Author: Nir Soffer <nsoffer@redhat.com>
Date:   Mon Nov 4 22:08:45 2013 +0200

    lvm: Do not use udev cache for obtaining device list

We removed the setting in RHEL 7:

commit e820cc58d0f89480ed4ce701c492501539a0b7ba
Author: Fred Rolland <frolland@redhat.com>
Date:   Sun Oct 18 14:51:31 2015 +0300

    lvm: Use udev cache for obtaining device list

We could not reproduce any error in RHEL 7.8, so looks like this broke
again in RHEL 8.2.

[1] https://bugzilla.redhat.com/1812801#c3
[2] https://gerrit.ovirt.org/c/109343/

Change-Id: I32b27685db808413144dfca59f3eaad296a22c8b
Bug-Url: https://bugzilla.redhat.com/1842053
Signed-off-by: Nir Soffer <nsoffer@redhat.com>