tests: Test convert to qcow2 compat=0.10
When converting to the old qcow2 compat=0.10 format without a backing
file, qemu-img allocates the entire image. Reproduce the issue in the
tests and mark the compat=0.10 as xfail.

There are 2 use cases:
- Convert base volume - we had a test for this, but we tested only
  compat=1.1. Add parameter for testing also compat=0.10.
- Collapse chain to target qcow2. Add new test for this use case testing
  both compat=1.1 and compat=0.10.

The failing tests succeed on Jeknins, I think because tests are using
NFS v3 storage, which does not support sparseness and breaks qemu-img
compare strict mode. Use userstorge mount with 512 and 4k sector size to
get consistent results.

Change-Id: I9d0fe63ea57c2d3f9a7738b6aef442242de02a03
Related-To: https://bugzilla.redhat.com/1858632
Signed-off-by: Nir Soffer <nsoffer@redhat.com>