tests: Add failing tests for copy_data
When converting to qcow2 compat=0.10 qemu allocates the entire image.
The current tests missed this since they checked only some data instead
of entire image.

Using qemu-img compare -s, we verify both content and allocation of the
destination image. This expose the failure when converting to older
domains using compat=0.10.

To mark only 2 permutations as failure we need to convert the test to
pytest and move it out of the class since pytest parameters do not work
in unittest.TestCase subclass.

The test for collapsing image was broken, always testing qcow2
destination image. The error was discovered by comparing with
qemuimg.comare() with specifying the image format.

When testing collapse, strict mode does not work, but we can compare
source and destination actual size to detect wrong allocation.

Finally to get consistent results with qemu-img compare strict mode, use
userstorage. Since we tests here also block storage and it is not
compatible with 4k storage, use storage with sector size of 512 bytes.

Change-Id: I826f4d2ce27537f1ced00e3b5f31f9efaa55ac03
Signed-off-by: Nir Soffer <nsoffer@redhat.com>